Grenade animation synched and grenades more reliable.

+++ network/slave_server/slave_server.cpp
@@ -200,7 +200,7 @@ void SlaveServer::clientConnect(uint32_t clientIndex)
 		int64_t currentTime = bx::getHPCounter();
 		const int64_t time = currentTime - m_startTime;
 		m_startTime = currentTime;
-		const float64 frequency = bx::getHPFrequency();
+		const float64 frequency = (float64)bx::getHPFrequency();
 		// seconds
 		float64 lastDeltaTime = float64(time * (1.0f / frequency));
 		m_totalTime += lastDeltaTime;
